requirements.txt
pytz==2022.7  # https://github.com/stub42/pytz
pillow==10.2.0  # https://github.com/python-pillow/Pillow
whitenoise==6.2.0  # https://github.com/evansd/whitenoise

# Django
# ------------------------------------------------------------------------------
django==4.0.8  # pyup: < 4.1  # https://www.djangoproject.com/
django-model-utils==4.3.1  # https://github.com/jazzband/django-model-utils
django-crispy-forms==1.14.0  # https://github.com/django-crispy-forms/django-crispy-forms
crispy-bootstrap5==0.7  # https://github.com/django-crispy-forms/crispy-bootstrap5
django-filter==23.5 # https://github.com/carltongibson/django-filter

# Django REST Framework
djangorestframework==3.14.0  # https://github.com/encode/django-rest-framework
django-cors-headers==3.13.0 # https://github.com/adamchainz/django-cors-headers


# Customize django admin
django-jet-reboot==1.3.5

# Environment variable
python-decouple==3.8

# messaging features
redis-cli
channels
daphne
channels-redis
django-redis


